+++ recipe/emailverification/api/generateEmailVerifyToken.go
@@ -34,7 +34,7 @@ func GenerateEmailVerifyToken(apiImplementation evmodels.APIInterface, options e
 
 	sessionContainer, err := session.GetSessionWithContext(
 		*options.Req, options.Res,
-		&sessmodels.GetSessionOptions{
+		&sessmodels.VerifySessionOptions{
 			OverrideGlobalClaimValidators: func(globalClaimValidators []claims.SessionClaimValidator, sessionContainer sessmodels.SessionContainer, userContext supertokens.UserContext) ([]claims.SessionClaimValidator, error) {
 				validators := []claims.SessionClaimValidator{}
 				return validators, nil