+++ recipe/session/setClaimValue_test.go
@@ -68,11 +68,11 @@ func TestShouldSetClaimValueOverwriteClaimValue(t *testing.T) {
 				Override: &sessmodels.OverrideStruct{
 					Functions: func(originalImplementation sessmodels.RecipeInterface) sessmodels.RecipeInterface {
 						oCreateNewSession := *originalImplementation.CreateNewSession
-						nCreateNewSession := func(userID string, accessTokenPayload map[string]interface{}, sessionDataInDatabase map[string]interface{}, disableAntiCsrf *bool, userContext supertokens.UserContext) (sessmodels.CreateNewSessionResponse, error) {
+						nCreateNewSession := func(userID string, accessTokenPayload map[string]interface{}, sessionDataInDatabase map[string]interface{}, disableAntiCsrf *bool, userContext supertokens.UserContext) (sessmodels.SessionContainer, error) {
 							trueClaim, _ := TrueClaim()
 							accessTokenPayload, err := trueClaim.Build(userID, accessTokenPayload, userContext)
 							if err != nil {
-								return sessmodels.CreateNewSessionResponse{}, err
+								return nil, err
 							}
 							return oCreateNewSession(userID, accessTokenPayload, sessionDataInDatabase, disableAntiCsrf, userContext)
 						}
@@ -126,11 +126,11 @@ func TestShouldSetClaimValueOverwriteClaimValueUsingHandle(t *testing.T) {
 				Override: &sessmodels.OverrideStruct{
 					Functions: func(originalImplementation sessmodels.RecipeInterface) sessmodels.RecipeInterface {
 						oCreateNewSession := *originalImplementation.CreateNewSession
-						nCreateNewSession := func(userID string, accessTokenPayload map[string]interface{}, sessionDataInDatabase map[string]interface{}, disableAntiCsrf *bool, userContext supertokens.UserContext) (sessmodels.CreateNewSessionResponse, error) {
+						nCreateNewSession := func(userID string, accessTokenPayload map[string]interface{}, sessionDataInDatabase map[string]interface{}, disableAntiCsrf *bool, userContext supertokens.UserContext) (sessmodels.SessionContainer, error) {
 							trueClaim, _ := TrueClaim()
 							accessTokenPayload, err := trueClaim.Build(userID, accessTokenPayload, userContext)
 							if err != nil {
-								return sessmodels.CreateNewSessionResponse{}, err
+								return nil, err
 							}
 							return oCreateNewSession(userID, accessTokenPayload, sessionDataInDatabase, disableAntiCsrf, userContext)
 						}
@@ -165,7 +165,7 @@ func TestShouldSetClaimValueOverwriteClaimValueUsingHandle(t *testing.T) {
 
 	sessInfo, err := GetSessionInformation(sessionContainer.GetHandle())
 	assert.NoError(t, err)
-	accessTokenPayload = sessInfo.AccessTokenPayload
+	accessTokenPayload = sessInfo.CustomClaimsInAccessTokenPayload
 	assert.Equal(t, 1, len(accessTokenPayload))
 	assert.Equal(t, false, accessTokenPayload["st-true"].(map[string]interface{})["v"])
 }